pyproject.toml
[tool.poetry]
name = "test-to-inforce"
version = "0.1.0"
description = ""
authors = ["marik <letschuka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
djoser = "^2.2.3"
python-decouple = "^3.8"
drf-yasg = "^1.21.7"

[tool.poetry.dev-dependencies]
pytest = "^8.2.2"
pytest-django = "^4.8.0"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
